Modify 6-size.c

+++ 0x00-hello_world/6-size.c
@@ -7,10 +7,10 @@
  */
 int main(void)
 {
-	printf("Size of a char: %lu byte(s)\n", sizeof(char));
-	printf("Size of an int: %lu byte(s)\n", sizeof(int));
-	printf("Size of a long: %lu byte(s)\n", sizeof(long));
-	printf("Size of a long long int: %lu byte(s)\n", sizeof(long long int));
-	printf("Size of a float: %lu byte(s\n)", sizeof(float));
+	printf("Size of a char: %d byte(s)\n", (int) sizeof(char));
+	printf("Size of an int: %d byte(s)\n", (int) sizeof(int));
+	printf("Size of a long: %d byte(s)\n", (int) sizeof(long));
+	printf("Size of a long long int: %d byte(s)\n", (int) sizeof(long long int));
+	printf("Size of a float: %d byte(s\n)", (int) sizeof(float));
 	return (0);
 }
